Predicted Score Block Coeff Stashing
Suggested work to move the coefficients from a bareblock of code into a table for safekeeping, to prevent an errant find-replace from introducing errors into the calculation unexpectedly. This commit adds a new table that stores the coefficients, and a block of pivot sql to utilize that table of coefficients to calculate the predicted score block.