Optimization - refactor with all suggestions from posted issues
This commit is a full refactor of the LoyaltyCohort script to include all the proposed optimization strategies. I have included an optional block, commented out, that adds the Charlson comorbidity index to the process. There are a few lines to uncomment in the final summary query as well to turn these fields on. Without charlson, the process completes in ~50s in UKY production environment. Charlson adds ~1m to the process.