Refactor DX code subquery for local site specific codes
Reduced complexity of the join in an attempt to make this more standard. The original code required sites to have dumped all the ACT Ontology tables into a single table called i2b2metadata (or a view). In the end I don't think it's necessary since the paths/concept_cd are already combined in the concept_dimension table. We can just join between it and the paths table. 

However, during UKY testing of this for our local BMI codes, it unavoidably inflates the num_dx1 & num_dx2 variables. We followed the example of MGB in the paths table and entered our local BMI paths as feature_name 'DX'. Our local BMI facts are not technically diagnoses, but actual calculated chart values that we convert to some bmi buckets. The team will need to discuss what to do about this. My instinct is to set the feature_name to something else like 'bmi', or 'custom', etc if it is NOT the icd9/10 BMI codes.