Change Cohort Flags to look for code_type SITE
Continuing from the last commit in regards to BMI as a vitals fact (not a DX), this seems like the easiest way to approach that type of coding situation. The end-users adding custom codes to the paths table will be responsible for ensuring that the code paths they add are not DX/PX/labs, and name them code_type = 'SITE' specifically when necessary. This may only effect BMI codes at most sites.