Add additional steps that make it possible to filter by the PSCutoff
This was most easily performed by adding an additional step to pre-calculate the agegrp before the final calculation step, instead of relying on grouping sets. This only adds about 2s to execution time on our small set. It does rely on a temp table, so a site with a large #cohort table will have a different experience. We technically need an extra dummy row for each patient in the All Patients group too, to join to the predictive score ntile CTE so we can filter the cohorts in the final calculation.