Refactor performance for consistence among sites; address 28&29
This refactor is difficult to parcel out in smaller commits due to the dependencies throughout the script. 

However, it does successfully address #28 by adding several parameters to allow a single procedure to do any amount of years lookback, as well as a parameter to force gendered numerators/denominators into the output. It addresses #29 by calculating the final charlson stats for the whole population, and the filtered top quintile separately. 

It exports the output to fit the spreadsheet created by @jklann with some additions for the new parameters, and a new parameter for SITE and EXTRACT_DDTM will help us ensure the output does not get mixed up in the spreadsheet. I've also included a column in the output for the runtime of the whole script so that we can keep an eye on performance across participating sites. 

One final thing to note on the new output: When the row is not filtered by the predicated score cutoff (CUTOFF_FILTER_YN='N') I have forced the predicted score cutoff to null. This made sense in my head, as the PCS was not used for these stats, and as the result set is very wide, if you can no longer see the CUTOFF_FILTER_YN variable, this could be helpful. It's a simple fix to flip this output back on later if we want to include it.