Adds two parameters to allow users to filter the LC by existing cohorts
In order to avoid running the procedure multiple times for each desired cohort, a new table variable parameter (user-defined type) was added. This allows the user to prepare and run ANY number of cohorts by passing the patient_num, cohort_name key pairs from outside the procedure. This functionality was the best approach to allow the loyalty cohort to be filtered by the new 4CE X.2 analysis cohorts, of which there are more than 30 different cohorts a patient can belong to.