Refactor to remove pain points
This commit is a refactor of the script to alleviate some codebase maintenance woes. We have decided to ALWAYS require the user to pass the cohort filter table, and to either fill in the index_dt in that table with their chosen static date, or the varied dates at the patient level. This alleviates the need for several parameters, and conditional statements throughout the script, and will make maintaining the codebase in the future much simpler.