Changes from working meeting 20220421
This commit includes the changes proposed and agreed on during a working meeting on the afternoon of 20220421. 

The procedure should now store permanently in the summary table the PercentPopulation window calculation, removing the need to store TotalSubjects in the 'PercentOfSubjects' records. This field is now forced to NULL after the calculation is performed. Our hope is to remove the potential for a site sharing "Percentage" data with another site accidentally sharing a non-obfuscated count of subjects in their output. 

More debug messages have been added throughout to keep the end-user informed of the progress of the script. 

The storage of loyalty_dev and loyalty_charlson_dev has been rewritten to be less destructive, and only remove records from those targets that match the CURRENT parameters being passed to the procedure. This should allow sites to store ALL runs of the script in these local analytic tables for their internal use (not just the most recent run of the procedure).