Contains a fix to PSC calculation by cohort AND agegrp
This step was previously only calculating the ntile for PCS by agegrp, ignoring cohort. Now that the query can ingest an arbitrary number of cohorts simultaneously, the ntile over statement needs to include cohort_name as well. 

There are some additional typo fixes. And minor fix to an oversight for storing the [SITE] code during the step that stores loyalty_dev and loyalty_cohort_dev.