DEATH_DT changes
This alteration comments out the step that deleted patients with a non-null death_dt. A new column for death_dt is added in the loyalty_dev and loyalty_charlson_dev tables. This can be leveraged by a site during an analysis step to handle deceased patient's as that analysis requires.