Alters date comparisons throughout
Optional change, but has the potential for significant performance gains. Alters join predicates throughout to avoid converting the start_date present in observation_fact or visit_dimension, and instead explicitly cast the date fields in the search to datetimes. The predicate rightly handles the conversion to datetime leaving a 00:00:00 timestamp by adding one day and using a less than comparison.